Fixed board fault highlighting long runtime
Fixed board fault highlighting long runtime. The problem in this case wasnt the copymakeborder function or the paste function. The image saving after each iteration was using up too much resources. The paster function doesnt require saving after every iteration so the save image option was moved to the end of the code.